requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.31.0
pydantic>=2.6.0
pillow>=10.0.0
pytest>=7.0.0
